CE move to ui
