+++ src/components/ui/CourseEnrollment.jsx
